chore: Move font utils to new Font module
