+++ mobile/src/components/Form/Input.tsx
@@ -7,7 +7,8 @@ import { usePreferenceStore } from "~/stores/Preference/store";
 
 import { FontSize } from "~/constants/Styles";
 import { OnRTL } from "~/lib/react";
-import { cn, getFont } from "~/lib/style";
+import { cn } from "~/lib/style";
+import { getFont } from "~/modules/font/utils";
 
 export function useInputRef() {
   return useRef<RNTextInput | RNGHTextInput>(null);
