+++ mobile/src/components/Typography/AccentText.tsx
@@ -3,7 +3,8 @@ import { Text } from "react-native";
 
 import { usePreferenceStore } from "~/stores/Preference/store";
 
-import { cn, getFont } from "~/lib/style";
+import { cn } from "~/lib/style";
+import { getFont } from "~/modules/font/utils";
 
 /** `<Text />` that utilizes the accent font. */
 export function AccentText({
