+++ mobile/src/lib/style.ts
@@ -2,9 +2,8 @@ import type { ClassValue } from "clsx";
 import { clsx } from "clsx";
 import { extendTailwindMerge } from "tailwind-merge";
 
-import { BorderRadius, FontFamily, FontSize } from "~/constants/Styles";
-import { toLowerCase } from "~/utils/string";
-import type { BundledFont } from "~/modules/font/constants";
+import { BorderRadius, FontSize } from "~/constants/Styles";
+import { FontFamily } from "~/modules/font/constants";
 import { ColorRoleOptions } from "~/modules/theme/constants";
 
 //#region Color
@@ -27,28 +26,6 @@ const customTwMerge = extendTailwindMerge({
 /** Combines any number of Tailwind classes nicely. */
 export function cn(...inputs: ClassValue[]) {
   return customTwMerge(clsx(inputs));
-}
-//#endregion
-
-//#region Font
-/**
- * Returns the correct font used from the codes used to determine the
- * accent & primary font used.
- */
-export function getFont(
-  font: BundledFont,
-  options?: { bold?: boolean; headline?: boolean },
-) {
-  const fontCode = font === "Geist Mono" ? "geistMono" : toLowerCase(font);
-  if ((options?.headline || options?.bold) && fontCode === "ntype") {
-    return FontFamily.ntypeHeadline;
-  } else if (
-    options?.bold &&
-    (fontCode === "geistMono" || fontCode === "roboto" || fontCode === "inter")
-  ) {
-    return FontFamily[`${fontCode}Medium`];
-  }
-  return FontFamily[fontCode];
 }
 //#endregion
