+++ mobile/src/modules/font/constants.ts
@@ -1,4 +1,18 @@
 //#region Bundled Fonts
+export const FontFamily = {
+  roboto: "Roboto-Regular",
+  robotoMedium: "Roboto-Medium",
+  inter: "Inter-Regular",
+  interMedium: "Inter-Medium",
+  geistMono: "GeistMono-Regular",
+  geistMonoMedium: "GeistMono-Medium",
+  ndot: "Ndot-77_JP_Extended",
+  ntype: "NType82-Regular",
+  ntypeHeadline: "NType82-Headline",
+
+  system: "System",
+} as const;
+
 export const BundledFontOptions = [
   "NDot",
   "NType",
