+++ mobile/src/navigation/screens/settings/AppUpdateView.tsx
@@ -11,10 +11,10 @@ import { useHasNewUpdate } from "~/navigation/hooks/useHasNewUpdate";
 import { ListLayout } from "~/navigation/layouts/ListLayout";
 
 import { FontSize } from "~/constants/Styles";
-import { getFont } from "~/lib/style";
 import { Links, openLink } from "~/lib/web-browser";
 import { SegmentedList } from "~/components/List/Segmented";
 import { AccentText } from "~/components/Typography/AccentText";
+import { getFont } from "~/modules/font/utils";
 import { useTheme } from "~/modules/theme/hooks";
 
 export default function AppUpdate() {
